+++ aurelix/crud/routes.py
@@ -1,4 +1,5 @@
-from fastapi import FastAPI, Request, HTTPException, Body
+from fastapi import FastAPI, Request, HTTPException, Body, UploadFile
+from fastapi.responses import RedirectResponse
 from fastapi.exceptions import ValidationException
 import json
 import pydantic
@@ -15,7 +16,7 @@ from ..utils import snake_to_pascal, snake_to_human, item_json
 
 
 def register_collection(app, Collection: type[BaseCollection], create_enabled=True, read_enabled=True, 
-                        update_enabled=True, delete_enabled=True, listing_enabled=True, 
+                        update_enabled=True, delete_enabled=True, listing_enabled=True, upload_enabled=True,
                         openapi_extra=None, max_page_size=100):
 
     openapi_extra = openapi_extra or {}
@@ -150,6 +151,21 @@ def register_collection(app, Collection: type[BaseCollection], create_enabled=Tr
                 }
             raise HTTPException(status_code=422, detail='Not Deleted')
         
+    if upload_enabled:
+        @Collection.view('/{identifier}/file/{field}', method='PUT', openapi_extra=openapi_extra, 
+                        summary='Get presigned url to upload file to %s' % snake_to_human(collection_name))
+        async def presigned_upload(request: Request, col: Collection, model: Model, identifier: str, field: str) -> RedirectResponse:
+            url = await col.get_presigned_upload_url(identifier, field)
+            print(url)
+            return RedirectResponse(url)
+
+        @Collection.view('/{identifier}/file/{field}', method='GET', openapi_extra=openapi_extra, 
+                        summary='Download file from %s' % snake_to_human(collection_name))
+        async def download(request: Request, col: Collection, model: Model, identifier: str, field: str) -> RedirectResponse:
+            url = await col.get_presigned_download_url(identifier, field)
+            return RedirectResponse(url)
+        
+
     if hasattr(Collection, 'StateMachine'):
 
         ModelTransition = pydantic.create_model(Schema.__name__ + 'Transition',